Fix compile error by moving types shared with typescript-deno-plugin to shared_types.d.ts
