+++ typescript-deno-plugin/src/index.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import type { PluginSettings, Settings } from "../../client/src/types";
+import type { PluginSettings, Settings } from "../../client/src/shared_types";
 import type * as ts from "../node_modules/typescript/lib/tsserverlibrary";
 
 /** Extract the return type from a maybe function. */
